+++ src/PackSite.Library.Logging.Serilog/SerilogConfigurationExtensions.cs
@@ -4,7 +4,6 @@
     using global::Serilog;
     using Microsoft.Extensions.Configuration;
     using Microsoft.Extensions.Hosting;
-    using PackSite.Library.Logging.Serilog.Internal;
 
     /// <summary>
     /// Serilog configuration extensions.
@@ -14,11 +13,7 @@
         /// <summary>
         /// Adds Serilog logging to appliation with logger options from app configuration.
         /// Additionally, the following enrichers are always set:
-        /// "App" property with appliaction name,
-        /// "Env" property with environment name,
-        /// "Ver" property with entry assembly version,
-        /// "FromLogContext", and
-        /// "WithExceptionDetails".
+        /// "FromLogContext".
         ///
         /// Optional enrichers:
         /// "WithMemoryUsage",
@@ -32,10 +27,20 @@
         /// </summary>
         /// <param name="builder"></param>
         /// <param name="configurationSectionName"></param>
+        /// <param name="writeToProviders">
+        /// By default, Serilog does not write events to Microsoft.Extensions.Logging.ILoggerProviders
+        /// registered through the Microsoft.Extensions.Logging API. Normally, equivalent
+        /// Serilog sinks are used in place of providers. Specify true to write events to all providers.
+        /// </param>
         /// <returns></returns>
-        public static IHostBuilder UseLogging(this IHostBuilder builder, string configurationSectionName = "Serilog")
+        [Obsolete(
+            $"This extension will be removed in PackSite.Library.Logging 2.0.0, " +
+            $"please use {nameof(SerilogLoggerConfigurationExtensions)}.{nameof(SerilogLoggerConfigurationExtensions.ConfigureWithFailSafeDefaults)}(...) instead.")]
+        public static IHostBuilder UseLogging(this IHostBuilder builder,
+                                              string configurationSectionName = "Serilog",
+                                              bool writeToProviders = false)
         {
-            return builder.UseLogging((context, services, loggerConfiguration) => { }, configurationSectionName);
+            return builder.UseLogging((context, services, loggerConfiguration) => { }, configurationSectionName, writeToProviders);
         }
 
         /// <summary>
@@ -56,10 +61,24 @@
         /// <param name="builder"></param>
         /// <param name="configureLogger"></param>
         /// <param name="configurationSectionName"></param>
+        /// <param name="writeToProviders">
+        /// By default, Serilog does not write events to Microsoft.Extensions.Logging.ILoggerProviders
+        /// registered through the Microsoft.Extensions.Logging API. Normally, equivalent
+        /// Serilog sinks are used in place of providers. Specify true to write events to all providers.
+        /// </param>
         /// <returns></returns>
-        public static IHostBuilder UseLogging(this IHostBuilder builder, Action<HostBuilderContext, LoggerConfiguration> configureLogger, string configurationSectionName = "Serilog")
+        [Obsolete(
+            $"This extension will be removed in PackSite.Library.Logging 2.0.0, " +
+            $"please use {nameof(SerilogLoggerConfigurationExtensions)}.{nameof(SerilogLoggerConfigurationExtensions.ConfigureWithFailSafeDefaults)}(...) instead.")]
+        public static IHostBuilder UseLogging(this IHostBuilder builder,
+                                              Action<HostBuilderContext, LoggerConfiguration> configureLogger,
+                                              string configurationSectionName = "Serilog",
+                                              bool writeToProviders = false)
         {
-            return builder.UseLogging((context, services, loggerConfiguration) => configureLogger(context, loggerConfiguration), configurationSectionName);
+            return builder.UseLogging((context, services, loggerConfiguration) =>
+            {
+                configureLogger(context, loggerConfiguration);
+            }, configurationSectionName, writeToProviders);
         }
 
         /// <summary>
@@ -80,17 +99,28 @@
         /// <param name="builder"></param>
         /// <param name="configureLogger"></param>
         /// <param name="configurationSectionName"></param>
+        /// <param name="writeToProviders">
+        /// By default, Serilog does not write events to Microsoft.Extensions.Logging.ILoggerProviders
+        /// registered through the Microsoft.Extensions.Logging API. Normally, equivalent
+        /// Serilog sinks are used in place of providers. Specify true to write events to all providers.
+        /// </param>
         /// <returns></returns>
-        public static IHostBuilder UseLogging(this IHostBuilder builder, Action<HostBuilderContext, IServiceProvider, LoggerConfiguration> configureLogger, string configurationSectionName = "Serilog")
+        [Obsolete(
+            $"This extension will be removed in PackSite.Library.Logging 2.0.0, " +
+            $"please use {nameof(SerilogLoggerConfigurationExtensions)}.{nameof(SerilogLoggerConfigurationExtensions.ConfigureWithFailSafeDefaults)}(...) instead.")]
+        public static IHostBuilder UseLogging(this IHostBuilder builder,
+                                              Action<HostBuilderContext, IServiceProvider, LoggerConfiguration> configureLogger,
+                                              string configurationSectionName = "Serilog",
+                                              bool writeToProviders = false)
         {
             builder.UseSerilog((context, services, loggerConfiguration) =>
             {
                 IConfiguration configuration = context.Configuration;
                 IHostEnvironment environment = context.HostingEnvironment;
 
-                loggerConfiguration.ConfigureSerilogCommons(configuration, configurationSectionName);
+                loggerConfiguration.ConfigureWithFailSafeDefaults(configuration, configurationSectionName);
                 configureLogger(context, services, loggerConfiguration);
-            }, false, false);
+            }, false, writeToProviders);
 
             return builder;
         }