+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+### v1.1.0 (21-Feb-2022)
+
+- Add new public extension `SerilogLoggerConfigurationExtensions.ConfigureWithFailSafeDefaults(...)`.
+- Depreciated `SerilogConfigurationExtensions.UseLogging`.
+
 ### v1.0.0 (09-Dec-2021)
 
 - First release.